Uploading an image to s3
Summary: Uploading an image to s3 given a path to the image

Test Plan:
patch to project
Verify that you have your AWS credentials set
navigate to src/js/ and open the file 'uploadImageTos3.js' with a text editor
NOTE: on line 7 I have the path to my credenatials. If yours if different please change to your directory.

Log in to AWS console.
Go to s3 buckets
select bucket testing-uswest2
notice that there is a file named, "testImage2.jpg"
remember the highest number on the file name. In this case it is 2 but after multiple tests the number should increase.

On line 10, update the number to the next. The idea is to keep the name unique.
For example, if the highest number is 2, then update this line to say 	testImage3.jpg

save file
npm install
node uploadImageTos3.js

refresh bucket. You should see your file now.

Click on the file name
click on open/download
open the downloaded file
you should be able to see a picture!

Reviewers: araceli19, austinz

Reviewed By: araceli19, austinz

Subscribers: austinz

Differential Revision: https://imagengine.phacility.com/D10